+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## v0.8.x
+
+### 💥️ Breaking Change
+
+- nothing
+
+### ✨ Feature
+
+- nothing
+
+### 🎨 Enhancement
+
+- nothing
+
+### 🐛 Fix
+
+- nothing
+
 ## v0.7.3
 
 ### 💥️ Breaking Change